life_management_agency/requirements.txt
python-dotenv
openai
agency-swarm
websockets
aiohttp
gradio
